+++ lib/libucsi/dvb/int_section.c
@@ -26,9 +26,8 @@ struct dvb_int_section * dvb_int_section_codec(struct section_ext *ext)
 {
 	uint8_t *buf = (uint8_t *) ext;
 	struct dvb_int_section *in = (struct dvb_int_section *) ext;
-
-	unsigned int pos = sizeof(struct section_ext);
-	unsigned int len = section_ext_length(ext);
+	size_t pos = sizeof(struct section_ext);
+	size_t len = section_ext_length(ext);
 
 	if (len < sizeof(struct dvb_int_section))
 		return NULL;
