+++ lib/libucsi/mpeg/pat_section.c
@@ -24,8 +24,8 @@
 struct mpeg_pat_section *mpeg_pat_section_codec(struct section_ext * ext)
 {
 	uint8_t *buf = (uint8_t *)ext;
-	int pos = sizeof(struct section_ext);
-	int len = section_ext_length(ext);
+	size_t pos = sizeof(struct section_ext);
+	size_t len = section_ext_length(ext);
 
 	if (len < sizeof(struct mpeg_pat_section))
 		return NULL;
